+++ KalikoCMS.Engine/Core/PageType.cs
@@ -33,6 +33,7 @@ namespace KalikoCMS.Core {
         public string DisplayName { get; set; }
         public string PageTypeDescription { get; set; }
         public string PageTemplate { get; set; }
+        public string PreviewImage { get; set; }
         public bool ShowPublishDates { get; set; }
         public bool ShowSortOrder { get; set; }
         public bool ShowVisibleInMenu { get; set; }