+++ KalikoCMS.Engine/Data/AutoMapperConfiguration.cs
@@ -43,6 +43,7 @@ namespace KalikoCMS.Data {
             Mapper.CreateMap<PageTypeEntity, PageType>()
                 .ForMember(m => m.AllowedTypes, o => o.Ignore())
                 .ForMember(m => m.Type, o => o.Ignore())
+                .ForMember(m => m.PreviewImage, o => o.Ignore())
                 .ReverseMap();
 
             Mapper.CreateMap<PropertyItem, PropertyItem>();